requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
pyautogui==0.9.54
pynput==1.7.6
scipy==1.11.4
pillow==10.0.1
tkinter-tooltip==2.1.0
matplotlib==3.10.7
